Importing Assets
Imported assets from the standart Unity Assets pack.  Fixed some issues with the scripts since this asset pack is now obsolete and is not available in the asset store, therefore I had to download it from another source. It is used in the section of the course so that's why I'm importing it