A method to switch between weapons and some fixes
Created a method that enables the player to switch between weapons either using the scrollwheel or the number buttons on the keyboard. This either enables or disables the weapon in the scene(using a foreach loop based on the player Input). Placed specific scripts in folders(All weapon related scripts in Weapon Scripts folder for example) and fixed a small issue where if the player shoots and then switches to another weapon he wont be able to shoot immediately(although the fix can now be exploited by constantly cycling trough the weapons and shooting with them)